// types.ts
export interface SeasonalDataItem {
    name: string;
    icon: string;
}